requirements.txt
Django==1.11.28
django-extensions==2.2.6
djangorestframework==3.9.4
drf-yasg==1.17.0
whitenoise==3.2
gunicorn==19.10.0
Markdown==3.1.1
pytz==2019.3
six==1.14.0
typing==3.7.4.1
pdfminer==20140328
requests==2.22.0
fake-useragent==0.1.11
django-environ==0.4.5
dj-database-url==0.5.0
psycopg2==2.8.4